requirements.txt
python-telegram-bot
requests
openai-whisper
python-dotenv
scipy